Optimizations
Instead of using the sorted ByteString as a key, this counts the
frequency of each letter in a sentence. It then uses the resulting
Vector as a key. This is cheaper than sorting a ByteString.

The list of sorted words that is used for deduping is only computed at
the last stage.

Two prefilters are performed:
* with the input list of strings, empty lists are ignored
* after the grouping of elements, entries with 1 element or less are
dropped before going to the deduplication

The Hashable instance could use some love, it should be possible to work
on data wider than Word8 at that place.